requirements.txt
speechrecognition
pyttsx3
psutil
pyjokes
wikipedia
cryptography
google.generativeai